# docs/openapi.generated.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/auth/me":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/auth/physician/dashboard-access":{"get":{"operationId":"AuthController_checkPhysicianAccess","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/observations/glucose":{"post":{"operationId":"GlucoseController_ingestGlucose","summary":"Ingestão de medição de glicemia (Manual ou Dispositivo)","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestGlucoseDto"}}}},"responses":{"201":{"description":"Observação registrada com sucesso."},"400":{"description":"Payload inválido ou fora dos limites plausíveis."},"409":{"description":"Evento duplicado (Idempotência)."}},"tags":["Observações Clínicas"],"security":[{"bearer":[]}]},"get":{"operationId":"GlucoseController_history","summary":"Histórico de medições de glicemia de um paciente (mais recentes primeiro)","parameters":[{"name":"patientId","required":true,"in":"query","description":"UUID do paciente","schema":{"type":"string"}}],"responses":{"200":{"description":"Lista de observações de glicemia."}},"tags":["Observações Clínicas"],"security":[{"bearer":[]}]}},"/api/v1/alerts":{"get":{"operationId":"AlertsController_list","summary":"Fila de atendimento: alertas ativos ordenados por urgência (P0 no topo)","parameters":[],"responses":{"200":{"description":"Lista de alertas ativos."}},"tags":["Alertas Clínicos"],"security":[{"bearer":[]}]}},"/api/v1/alerts/{id}/assume":{"patch":{"operationId":"AlertsController_assume","summary":"Assumir (lock) um alerta OPEN","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Alerta assumido."},"409":{"description":"Alerta já está sendo tratado por outro profissional."}},"tags":["Alertas Clínicos"],"security":[{"bearer":[]}]}},"/api/v1/alerts/{id}/resolve":{"patch":{"operationId":"AlertsController_resolve","summary":"Resolver (fechar) um alerta assumido, com conduta clínica","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResolveAlertDto"}}}},"responses":{"200":{"description":"Alerta resolvido."},"409":{"description":"Alerta não está em atendimento pelo profissional."}},"tags":["Alertas Clínicos"],"security":[{"bearer":[]}]}},"/api/v1/analytics/population-risk":{"get":{"operationId":"AnalyticsController_getPopulationRisk","summary":"Estratificação Populacional por Tiers de Risco (US-21)","parameters":[],"responses":{"200":{"description":"Distribuição de risco retornada com sucesso."}},"tags":["Portal da Operadora - Analytics"],"security":[{"bearer":[]}]}},"/api/v1/analytics/care-gaps":{"get":{"operationId":"AnalyticsController_getCareGaps","summary":"Monitoramento de Lacunas de Cuidado Vencidas (US-21)","parameters":[],"responses":{"200":{"description":"Lista priorizada de lacunas de cuidado."}},"tags":["Portal da Operadora - Analytics"],"security":[{"bearer":[]}]}},"/api/v1/analytics/utilization-stats":{"get":{"operationId":"AnalyticsController_getUtilizationStats","summary":"Indicadores de Utilização Hospitalar / 1.000 Beneficiários (US-22)","parameters":[],"responses":{"200":{"description":"Taxas de PA e internações / 1.000 beneficiários."}},"tags":["Portal da Operadora - Analytics"],"security":[{"bearer":[]}]}},"/api/v1/operator/analytics/overview":{"get":{"operationId":"OperatorAnalyticsController_getOverview","summary":"Visão Geral Populacional de Risco e Lacunas (US-21 + US-22)","parameters":[],"responses":{"200":{"description":"Indicadores populacionais retornados com sucesso."}},"tags":["Portal da Operadora - Analytics Populacional"],"security":[{"bearer":[]}]}},"/api/v1/health":{"get":{"operationId":"HealthController_checkHealth","summary":"Verificação de Saúde do Serviço (Docker Healthcheck)","parameters":[],"responses":{"200":{"description":""}},"tags":["Healthcheck"]}}},"info":{"title":"Lifecode Platform API","description":"Especificação oficial dos contratos de serviços da plataforma de gestão do diabetes.","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"TenantHeader":{"type":"apiKey","in":"header","name":"X-Tenant-ID"}},"schemas":{"RegisterDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{}},"SymptomsReportedDto":{"type":"object","properties":{"confusionOrAlteredConsciousness":{"type":"boolean","description":"Confusão mental ou consciência alterada"},"sweatingOrTremors":{"type":"boolean","description":"Sudorese ou tremores"},"vomitingOrKetoneSigns":{"type":"boolean","description":"Vômitos ou sinais de cetose"}}},"IngestGlucoseDto":{"type":"object","properties":{"schemaVersion":{"type":"string","description":"Versão do schema","default":"1.0"},"patientId":{"type":"string","description":"UUID do paciente","format":"uuid"},"externalEventId":{"type":"string","description":"ID externo do evento para idempotência"},"value":{"type":"number","description":"Valor da glicemia (na unidade informada em \"unit\")"},"unit":{"type":"string","description":"Unidade de medida (UCUM)","enum":["mg/dL","mmol/L"],"default":"mg/dL"},"sourceType":{"type":"string","description":"Origem da medição","enum":["CGM","FLASH","CAPILLARY","MANUAL","LABORATORY"]},"context":{"type":"string","description":"Contexto da medição","enum":["FASTING","PRE_PRANDIAL","POST_PRANDIAL","BEDTIME","NOCTURNAL","RANDOM"]},"deviceId":{"type":"string","description":"ID do dispositivo"},"measuredAt":{"type":"string","description":"Data e hora da medição (ISO 8601 UTC)"},"symptomsReported":{"description":"Sintomas relatados pelo paciente","allOf":[{"$ref":"#/components/schemas/SymptomsReportedDto"}]},"notes":{"type":"string","description":"Observações adicionais","maxLength":500}},"required":["patientId","externalEventId","value","unit","sourceType","measuredAt"]},"ResolveAlertDto":{"type":"object","properties":{"disposition":{"type":"string","description":"Conduta clínica registrada no fechamento","enum":["PATIENT_CONTACTED_STABLE","EMERGENCY_SERVICES_DISPATCHED","ESCALATED_TO_PHYSICIAN","FALSE_ALARM","OTHER"]},"notes":{"type":"string","description":"Notas da conduta clínica","maxLength":1000}},"required":["disposition"]}}}}